Newest Version with Build
Comes with working highscore system! The highest score that I have achieved while testing is the current high score to beat. If you would like to change this then modify the "highScore" variable in the HighScoreScript. New button on main menu for high score. Will add notification noise/GUI in game if new high score is achieved. May be a bug with Ball material, will continue to check and update. New bugs have not been discovered, so other than the ball bug, game is fairly stable and optimized.